+++ src/main/java/com/hendisantika/sekolah/SekolahApplication.java
@@ -1,6 +1,6 @@
 package com.hendisantika.sekolah;
 
-import nz.net.ultraq.thymeleaf.LayoutDialect;
+import nz.net.ultraq.thymeleaf.layoutdialect.LayoutDialect;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.Bean;
