+++ src/main/java/com/hendisantika/sekolah/repository/AlbumRepository.java
@@ -2,6 +2,7 @@ package com.hendisantika.sekolah.repository;
 
 import com.hendisantika.sekolah.entity.Album;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
 /**
  * Created by IntelliJ IDEA.
@@ -12,5 +13,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
  * Date: 03/04/20
  * Time: 07.53
  */
+@Repository
 public interface AlbumRepository extends JpaRepository<Album, Long> {
 }
