+++ src/main/java/com/hendisantika/sekolah/repository/FilesRepository.java
@@ -2,6 +2,7 @@ package com.hendisantika.sekolah.repository;
 
 import com.hendisantika.sekolah.entity.Files;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
 import java.util.UUID;
 
@@ -14,5 +15,6 @@ import java.util.UUID;
  * Date: 18/03/20
  * Time: 18.31
  */
+@Repository
 public interface FilesRepository extends JpaRepository<Files, UUID> {
 }
