+++ src/main/java/com/hendisantika/sekolah/repository/InboxRepository.java
@@ -2,8 +2,10 @@ package com.hendisantika.sekolah.repository;
 
 import com.hendisantika.sekolah.entity.Inbox;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
 import java.util.UUID;
 
+@Repository
 public interface InboxRepository extends JpaRepository<Inbox, UUID> {
 }
