+++ src/main/java/com/hendisantika/sekolah/repository/KategoriRepository.java
@@ -2,6 +2,8 @@ package com.hendisantika.sekolah.repository;
 
 import com.hendisantika.sekolah.entity.Kategori;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.lang.NonNull;
+import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
@@ -14,6 +16,7 @@ import java.util.Optional;
  * Date: 18/03/20
  * Time: 14.23
  */
+@Repository
 public interface KategoriRepository extends JpaRepository<Kategori, Long> {
-    Optional<Kategori> findById(Long kategoriId);
+    @NonNull Optional<Kategori> findById(@NonNull Long kategoriId);
 }
