+++ src/main/java/com/hendisantika/sekolah/repository/KelasRepository.java
@@ -2,6 +2,8 @@ package com.hendisantika.sekolah.repository;
 
 import com.hendisantika.sekolah.entity.Kelas;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
+@Repository
 public interface KelasRepository extends JpaRepository<Kelas, Integer> {
 }
