+++ src/main/java/com/hendisantika/sekolah/service/PenggunaService.java
@@ -1,41 +1,13 @@
 package com.hendisantika.sekolah.service;
 
 import com.hendisantika.sekolah.entity.Pengguna;
-import com.hendisantika.sekolah.repository.PenggunaRepository;
-import org.springframework.stereotype.Service;
 
 import java.util.List;
 
-/**
- * Created by IntelliJ IDEA.
- * Project : sekolah
- * User: hendisantika
- * Email: hendisantika@gmail.com
- * Telegram : @hendisantika34
- * Date: 20/03/20
- * Time: 05.57
- */
-@Service
-public class PenggunaService {
-    private final PenggunaRepository penggunaRepository;
+public interface PenggunaService {
+    Pengguna saveUser(Pengguna user);
+    List<Pengguna> getUsers();
 
-    public PenggunaService(PenggunaRepository penggunaRepository) {
-        this.penggunaRepository = penggunaRepository;
-    }
+//    UserDetails loadUserByUsername(String username);
 
-    public Pengguna saveUser(Pengguna user) {
-        return penggunaRepository.save(user);
-    }
-
-    public List<Pengguna> getUsers() {
-        return penggunaRepository.findAll();
-    }
-
-//    public UserDetails loadUserByUsername(String username) {
-//        Pengguna user = penggunaRepository.findByUsername(username);
-//        if (user == null) {
-//            throw new UsernameNotFoundException(username);
-//        }
-//        return new MyUserPrincipal(user);
-//    }
 }
