+++ src/main/java/com/hendisantika/sekolah/service/impl/UserDetailsServiceImpl.java
@@ -9,6 +9,7 @@ import org.springframework.security.core.userdetails.User;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -23,6 +24,7 @@ import java.util.Optional;
  * Date: 21/03/20
  * Time: 06.28
  */
+@Service
 public class UserDetailsServiceImpl implements UserDetailsService {
 
     private final PenggunaRepository penggunaRepository;
