+++ src/main/resources/templates/fragments/dashboard/header.html
@@ -1,4 +1,4 @@
-<html xmlns:th="http://www.thymeleaf.org">
+<html xmlns:th="http://www.thymeleaf.org" lang="en">
 <head th:fragment="header">
     <meta charset="utf-8">
     <meta content="IE=edge" http-equiv="X-UA-Compatible">
