+++ src/main/resources/templates/fragments/dashboard/layout.html
@@ -1,6 +1,6 @@
 <!DOCTYPE html>
 <html lang="en" xmlns:th="http://www.thymeleaf.org">
-<head th:replace="fragments/dashboard/header :: header">
+<head th:replace="~{fragments/dashboard/header :: header}">
 </head>
 
 <body class="nav-md">
