+++ src/main/resources/templates/fragments/footer.html
@@ -1,5 +1,5 @@
 <!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
+<html xmlns:th="http://www.thymeleaf.org" lang="en">
 <body>
 <!--============================= FOOTER =============================-->
 <footer th:fragment="footer">
