+++ src/main/resources/templates/fragments/header.html
@@ -62,7 +62,7 @@
                     </button>
                     <a class="navbar-brand nav-brand2" href="/"><img class="img img-responsive"
                                                                      th:src="@{~/theme/images/logo-dark.png}"
-                                                                     width="200px;"></a>
+                                                                     width="200px;" alt=""></a>
                     <div class="collapse navbar-collapse justify-content-end" id="navbarNavDropdown">
                         <ul class="navbar-nav">
                             <li class="nav-item">
