+++ src/main/resources/templates/fragments/layout.html
@@ -1,14 +1,14 @@
 <!DOCTYPE html>
 <html lang="en" xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout" xmlns:th="http://www.thymeleaf.org">
-    <head th:replace="fragments/header :: head">
+    <head th:replace="~{fragments/header :: head}">
 
     <body>
         <!--============================= HEADER =============================-->
-        <div th:replace="fragments/header :: header"/>
-        <div th:replace="fragments/header :: menu"/>
-        <div layout:fragment="content"/>
-        <div th:replace="fragments/footer :: footer"/>
-        <div th:replace="fragments/footer :: scripts"/>
+        <div th:replace="~{fragments/header :: header}"></div>
+        <div th:replace="~{fragments/header :: menu}"></div>
+        <div layout:fragment="content"></div>
+        <div th:replace="~{fragments/footer :: footer}"></div>
+        <div th:replace="~{fragments/footer :: scripts}"></div>
 
     </body>
