+++ src/main/resources/templates/index.html
@@ -1,14 +1,14 @@
 <!DOCTYPE html>
 <html lang="en" layout:decorate="~{fragments/layout}" xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout"
       xmlns:th="http://www.thymeleaf.org">
-<head th:replace="fragments/header :: head"/>
+<head th:replace="~{fragments/header :: head}"></head>
 <body>
-<div th:replace="fragments/header :: header"/>
-<div th:replace="fragments/header :: menu"/>
+<div th:replace="~{fragments/header :: header}"></div>
+<div th:replace="~{fragments/header :: menu}"></div>
 <!--//END HEADER -->
 <!--============================= ABOUT =============================-->
 <div layout:fragment="content">
-    <section th:replace="fragments/header :: carousel"/>
+    <section th:replace="~{fragments/header :: carousel}"></section>
     <section class="clearfix about about-style2">
         <div class="container">
             <div class="row">
@@ -98,11 +98,11 @@
         </div>
     </div>
 </section>
-    <div class="detailed_chart" th:replace="fragments/header :: chart"/>
+    <div class="detailed_chart" th:replace="~{fragments/header :: chart}"></div>
 </div>
 <!--//END EVENTS -->
-<div th:replace="fragments/footer :: footer"/>
-<div th:replace="fragments/footer :: scripts"/>
+<div th:replace="~{fragments/footer :: footer}"></div>
+<div th:replace="~{fragments/footer :: scripts}"></div>
 </body>
 
 </html>
