+++ src/main/resources/templates/samples/index2.html
@@ -1,13 +1,12 @@
 <!DOCTYPE html>
 <html lang="en" layout:decorate="~{fragments/layout}" xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout"
       xmlns:th="http://www.thymeleaf.org">
-<head th:replace="fragments/header :: head"/>
-<div th:replace="fragments/header :: header"/>
+<head th:replace="~{fragments/header :: head}"></head>
+<div th:replace="~{fragments/header :: header}"></div>
 <body>
-<div th:replace="fragments/header :: menu"/>
+<div th:replace="~{fragments/header :: menu}"></div>
 
-<body>
-<section th:replace="fragments/header :: carousel"/>
+<section th:replace="~{fragments/header :: carousel}"></section>
 <!--//END HEADER -->
 <!--============================= ABOUT =============================-->
 <div th:fragment="content">
@@ -101,11 +100,10 @@
             </div>
         </div>
     </section>
-    <div class="detailed_chart" th:replace="fragments/header :: chart"/>
+    <div class="detailed_chart" th:replace="~{fragments/header :: chart}"></div>
 </div>
 <!--//END EVENTS -->
-<div th:replace="fragments/footer :: footer"/>
-<div th:replace="fragments/footer :: scripts"/>
+<div th:replace="~{fragments/footer :: footer}"></div>
+<div th:replace="~{fragments/footer :: scripts}"></div>
 </body>
-
 </html>
