+++ src/main/resources/templates/samples/upload.html
@@ -1,5 +1,5 @@
 <!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
+<html xmlns:th="http://www.thymeleaf.org" lang="en">
 <link rel="stylesheet" th:href="@{/css/default.css}"/>
 <head>
     <meta charset="ISO-8859-1"/>
