+++ src/main/resources/templates/siswa.html
@@ -1,10 +1,10 @@
 <!DOCTYPE html>
 <html lang="en" layout:decorate="~{fragments/layout}" xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout"
       xmlns:th="http://www.thymeleaf.org">
-<head th:replace="fragments/header :: head"/>
-<div th:replace="fragments/header :: header"/>
+<head th:replace="~{fragments/header :: head}"></head>
+<div th:replace="~{fragments/header :: header}"></div>
 <body>
-<div th:replace="fragments/header :: menu"/>
+<div th:replace="~{fragments/header :: menu}"></div>
 
 <div layout:fragment="content">
     <section class="our-teachers">
@@ -22,7 +22,8 @@
                         <img alt="#" class="img-fluid" th:src="*{'data:image/*;base64,'+ s.photoBase64}">
                         <p class="text-center mt-3"><span th:text="${s.nama}">nama</span>
                             <br>
-                            <span th:text="${s.kelas.nama}">Kelas</p>
+                            <span th:text="${s.kelas.nama}">Kelas</span>
+                        </p>
                     </div>
                 </div>
             </div>
@@ -32,7 +33,7 @@
     </section>
 </div>
 
-<div th:replace="fragments/footer :: footer"/>
-<div th:replace="fragments/footer :: scripts"/>
+<div th:replace="~{fragments/footer :: footer}"></div>
+<div th:replace="~{fragments/footer :: scripts}"></div>
 </body>
 </html>
